requirements.txt
amqp==5.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
boto3==1.36.2
botocore==1.36.2
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.6
django-chunked-upload==2.0.0
django-cors==0.1
django-cors-headers==4.6.0
django-phonenumber-field==8.0.0
django-shortuuidfield==0.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flower==2.0.1
google==3.0.0
google-api-core==2.24.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
httplib2==0.22.0
humanize==4.11.0
idna==3.10
jmespath==1.0.1
kombu==5.3.7
oauthlib==3.2.2
phonenumbers==8.13.40
pillow==10.4.0
prometheus_client==0.21.1
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.29.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyJWT==2.8.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.1
shortuuid==1.0.13
six==1.16.0
soupsieve==2.6
sqlparse==0.5.0
tornado==6.4.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
django-storages
watchdog